# Conflicts:
#	ruoyi-client/src/main/java/com/ruoyi/client/controller/MatchRankingController.java
#	ruoyi-client/src/main/java/com/ruoyi/client/controller/ProjectController.java
#	ruoyi-client/src/main/java/com/ruoyi/client/domain/entity/Project.java
#	ruoyi-client/src/main/java/com/ruoyi/client/domain/vo/MatchPankingVO.java
#	ruoyi-client/src/main/java/com/ruoyi/client/mapper/MatchPankingMapper.java
#	ruoyi-client/src/main/java/com/ruoyi/client/mapper/ProjectMapper.java
#	ruoyi-client/src/main/java/com/ruoyi/client/service/MatchRankingService.java
#	ruoyi-client/src/main/java/com/ruoyi/client/service/ProjectService.java
#	ruoyi-client/src/main/java/com/ruoyi/client/service/impl/MatchPankingServiceImpl.java
#	ruoyi-client/src/main/java/com/ruoyi/client/service/impl/ProjectServiceImpl.java
#	ruoyi-client/src/main/resources/mapper/MatchPankingMapper.xml
#	ruoyi-client/src/main/resources/mapper/ProjectMapper.xml
#	ruoyi-client/src/main/resources/templates/match/matchRanking.html
#	ruoyi-client/src/main/resources/templates/sports/sports.html
#	ruoyi-client/src/main/resources/templates/user/index.html
#	ruoyi-client/src/main/resources/templates/user/login.html

+++ ruoyi-client/src/main/java/com/ruoyi/client/domain/dto/UserUpdateDTO.java
@@ -7,9 +7,9 @@ import lombok.Data;
  */
 @Data
 public class UserUpdateDTO {
-    private Integer userId;
+    private Long userId;
     private String type;
-    private Integer typeId;
+    private Long typeId;
     private String changeType;
     private String value;
     private String token;
