+++ ruoyi-client/src/main/java/com/ruoyi/client/mapper/UserMapper.java
@@ -30,12 +30,12 @@ public interface UserMapper extends BaseMapper<User> {
      * @param userTypeStudent 用户类型
      * @return 用户数据
      */
-    User selectUserByType(@Param("studentId") Integer studentId,@Param("userTypeStudent") String userTypeStudent);
+    User selectUserByType(@Param("studentId") Long studentId, @Param("userTypeStudent") String userTypeStudent);
 
     /**
      * 查找学生信息
      * @param userId 学生id
      * @return 学生信息
      */
-    UserInfoVO selectStudentInfo(Integer userId);
+    UserInfoVO selectStudentInfo(Long userId);
 }
